Refactor code structure: separate file I/O, data processing, and LLM analysis into distinct modules;
